LPS-79878 Inline class.
